trying to move function from view and model to controller
