+++ src/main/java/io/github/oliviercailloux/teach_spreadsheets/gui/Model.java
@@ -84,60 +84,4 @@ public class Model {
 	public Set<CoursePrefElement> getChosenPreferences() {
 		return chosenPreferences;
 	}
-
-	/**
-	 * Updates sets of CoursePrefElement thanks to the data retrieved from a table
-	 * item
-	 * 
-	 * @param stringTableItem the strings shown from the table item. Its size is 4 :
-	 *                        first element is teacher name, second is course name,
-	 *                        third is group type and fourth is teacher choice
-	 * @param source          the set where we want to find the CoursePrefElement
-	 *                        object corresponding to stringTableItem
-	 * @param target          the set where we want to add the CoursePrefElement
-	 *                        object corresponding to stringTableItem
-	 */
-	private void updateSet(List<String> stringTableItem, Set<CoursePrefElement> source,
-			Set<CoursePrefElement> target) {
-		checkNotNull(stringTableItem);
-		checkArgument(stringTableItem.size() == 4);
-		checkNotNull(source);
-		checkNotNull(target);
-
-		for (CoursePrefElement coursePrefElement : source) {
-
-			List<String> stringCoursePrefElement = coursePrefElement.getDataForTableItem();
-
-			if (stringCoursePrefElement.get(0).equals(stringTableItem.get(0))
-					&& stringCoursePrefElement.get(1).equals(stringTableItem.get(1))
-					&& stringCoursePrefElement.get(2).equals(stringTableItem.get(2))
-					&& stringCoursePrefElement.get(3).equals(stringTableItem.get(3))) {
-				source.remove(coursePrefElement);
-				target.add(coursePrefElement);
-				break;
-			}
-		}
-	}
-
-	/**
-	 * This method is called from Controller class when a table item has been
-	 * clicked. Updates the data from Model thanks to the table item data.
-	 * 
-	 * @param texts               the strings shown from the table item. texts' size
-	 *                            is 4 : first element is teacher name, second is
-	 *                            course name, third is group type and fourth is
-	 *                            teacher choice
-	 * @param toChosenPreferences true iff the element that has been clicked is on
-	 *                            the Table named all preferences
-	 */
-	public void updatePreferences(ArrayList<String> texts, boolean toChosenPreferences) {
-		checkNotNull(texts);
-		checkArgument(texts.size() == 4);
-
-		if (toChosenPreferences) {
-			updateSet(texts, allPreferences, chosenPreferences);
-		} else {
-			updateSet(texts, chosenPreferences, allPreferences);
-		}
-	}
 }
